% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{score_test}
\alias{score_test}
\title{Compute (modified) score test statistic in a linear mixed model}
\usage{
score_test(y, X, Z, Beta, sigma, lambda, lam_idx, test_idx)
}
\arguments{
\item{y}{a vector of observed responses}

\item{X}{a matrix of predictors whose i:th row corresponds to the i:th
element in y}

\item{Z}{a design matrix for the random effects whose i:th row corresponds
to the i:th element in y}

\item{Beta}{a vector of regression coefficients of length ncol(X)}

\item{sigma}{the standard deviation of the error term}

\item{lambda}{a vector of scale parameters (standard deviations) of the
random effects}

\item{lam_idx}{a vector of length ncol(Z) whose j:th element indicates
which element of lambda scales the j:th random effect}

\item{test_idx}{a vector of integers indicating for which elements of
theta = c(Beta, sigma, lambda) the test statistic is to be computed}
}
\value{
a vector with test-statistic ("chi_sq"), degrees of freedom ("df"),
and p-value ("p_val")
}
\description{
The statistic is based on a modified score function that
replaces the first partial derivative of the log-likelihood with
respect to a standard deviation parameter by the second partial
derivative at points where the former is almost surely equal to zero,
which happens if and only if the standard deviation parameter is equal
to zero. When no standard deviation parameters are zero under the null
hypothesis, the test statistic is the usual score test statistic
standardized by expected Fisher information.
}
